fix: add guard against user id not being present
If `process.env.AUTH_MODE` is set to `SKIP` then auth0 authentication
will be skipped and the user context not populated.
This leads to incorrect behaviour in cases that assume the user
information is present in the context.
This change adds assertions to these code sections so they will
fail.