fix: set correct ports for frontend and backend
This reverts 680dc6d5029fc7f16d49a135afe728f3cf1553a4

The backend server is configured to be running at port 3000
The frontend dev server is configured to be running at port 8080

The static file hosting is removed from the backend server,
this is handled by the frontend dev server.

The behaviour after this fix is already documented in the
frontend and backend READMEs.