fix: correctly mark onboarding as completed
This changes the implementation to fetch the
list of groups the user is captain of.
While the backend prevents the creation of more
than one group this implementation provides
some robustness against future changes.

The actual problem is that the backend
does not populate the groupId in the session
cookie, therefore it is always undefined.