refactor(group): make country and locality regular property
This converts the primaryLocation field of a group into two
regular, required fields: country and locality.

Locality is used as a more generic term instead of city.

Currently there is no secondary or other location, so
primary is not neccessary to depict the groups location.

Additionally, all groups can be expected to provide a
country and a locality to identify them in combination
with their group name.

Making country a regular field also simplifies the validation,
and we now also use the list of countries.